+++ editor/import/resource_importer_image.cpp
@@ -81,7 +81,7 @@ Error ResourceImporterImage::import(const String &p_source_file, const String &p
 		ERR_FAIL_COND_V(!f, ERR_CANT_OPEN);
 	}
 
-	size_t len = f->get_len();
+	int64_t len = f->get_len();
 
 	Vector<uint8_t> data;
 	data.resize(len);
